+++ frontend/src/components/Sidebar/Sidebar.module.css
@@ -8,6 +8,7 @@
   margin-left: 15px;
   top: 15px;
   font-size: 0.9rem;
+  margin-right: 0;
 }
 
 .sidebarHeader {
