+++ frontend/src/pages/Dashboard/Dashboard.module.css
@@ -1,0 +1,3 @@
+.dashboard{
+  display: flex;
+}
